+++ core/build.gradle.kts
@@ -26,14 +26,13 @@ plugins {
 // -----------------------------------------------------------------------------
 // Dependencies
 // -----------------------------------------------------------------------------
-
-val junitTestCompile = configurations.create("junitTestCompile")
-
 dependencies {
-    api(project(":knotx-fragments-engine-api"))
-}
+    api(project(":knotx-fragments-handler-api"))
+    implementation(project(":knotx-fragments-engine"))
 
-junitTestCompile.extendsFrom(configurations.named("testImplementation").get())
+    implementation(group = "org.apache.commons", name = "commons-lang3")
+    implementation(group = "io.vertx", name = "vertx-circuit-breaker")
+}
 
 // -----------------------------------------------------------------------------
 // Source sets
@@ -41,27 +40,13 @@ junitTestCompile.extendsFrom(configurations.named("testImplementation").get())
 sourceSets.named("main") {
     java.srcDir("src/main/generated")
 }
-tasks.named<JavaCompile>("compileJava") {
-    options.annotationProcessorGeneratedSourcesDirectory = file("src/main/generated")
-}
-tasks.named<Delete>("clean") {
-    delete.add("src/main/generated")
-}
-sourceSets.create("junitTest") {
-    compileClasspath += sourceSets.named("main").get().output
-}
-sourceSets.named("test") {
-    compileClasspath += sourceSets.named("junitTest").get().output
-    runtimeClasspath += sourceSets.named("junitTest").get().output
-}
-
 
 // -----------------------------------------------------------------------------
 // Tasks
 // -----------------------------------------------------------------------------
 tasks {
     named<RatTask>("rat") {
-        excludes.addAll("**/*.json", "**/*.md", "**/*.adoc", "**/build/*", "**/out/*", "**/generated/*", "/src/test/resources/*")
+        excludes.addAll("**/*.json", "**/*.md", "**/*.adoc", "**/build/*", "**/out/*", "**/generated/*")
     }
     getByName("build").dependsOn("rat")
 }
@@ -82,22 +67,16 @@ tasks.named<Javadoc>("javadoc") {
         (options as StandardJavadocDocletOptions).addBooleanOption("html5", true)
     }
 }
-tasks.register<Jar>("testJar") {
-    from(sourceSets.named("junitTest").get().output)
-    classifier = "tests"
-}
-
 publishing {
     publications {
         create<MavenPublication>("mavenJava") {
-            artifactId = "knotx-knot-engine-core"
+            artifactId = "knotx-knot-engine-handler"
             from(components["java"])
             artifact(tasks["sourcesJar"])
             artifact(tasks["javadocJar"])
-            artifact(tasks["testJar"])
             pom {
-                name.set("Knot.x Knot Engine Core")
-                description.set("Knot Engine Core module containing engine implementation.")
+                name.set("Knot.x Knot Engine Handler")
+                description.set("Knot Engine Handler allowing to configure engine in request processing chain.")
                 url.set("http://knotx.io")
                 licenses {
                     license {
@@ -143,9 +122,9 @@ publishing {
         }
     }
 }
-
 signing {
     sign(publishing.publications["mavenJava"])
 }
 
 apply(from = "../gradle/common.deps.gradle.kts")
+apply(from = "../gradle/codegen.deps.gradle.kts")
