+++ core/src/main/java/io/knotx/engine/handler/exception/ConfigurationException.java
