+++ core/src/main/java/io/knotx/engine/handler/exception/GraphConfigurationException.java
