+++ core/src/main/java/io/knotx/engine/handler/proxy/EBOperationProxy.java
