+++ core/src/main/resources/META-INF/services/io.knotx.server.api.handler.RoutingHandlerFactory
