+++ engine/assets/images/graph_node.png
