+++ engine/src/main/java/io/knotx/engine/core/EventLogEntry.java
