+++ engine/src/main/java/io/knotx/engine/core/GraphNode.java
